+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.1.0](https://github.com/Stacktical/stacktical-developer-toolkit/compare/v1.0.0...v1.1.0) (2021-05-25)
+
+
+### Features
+
+* chainlink utils to multi node ([586e1ef](https://github.com/Stacktical/stacktical-developer-toolkit/commit/586e1ef92201d7c95d7e4f9b3c99d7d36dfd2cd4))
+* creating compose file for prod chainlink nodes ([3d071ce](https://github.com/Stacktical/stacktical-developer-toolkit/commit/3d071ce5e28e28b4eafce89332581f1c2d74599a))
+* deploying multiple docker compose locally ([3464495](https://github.com/Stacktical/stacktical-developer-toolkit/commit/3464495b2d518f624d44ea7bfc7687dd23bd6918))
+* deploying to multiple chainlink nodes ([3dccef4](https://github.com/Stacktical/stacktical-developer-toolkit/commit/3dccef401a92fab86d1a265d45d0c2dccc8e51f5))
+* deployment to harmonytestnet automated ([f034292](https://github.com/Stacktical/stacktical-developer-toolkit/commit/f0342923e40ec071b91da73cbbe0978e86960649))
+* getting service agreement configuration ([cee0ae1](https://github.com/Stacktical/stacktical-developer-toolkit/commit/cee0ae1a92d52ef1b7fc02279b427123752ffaea))
+* harmonytestnet hardhat configuration ([888b474](https://github.com/Stacktical/stacktical-developer-toolkit/commit/888b47462c23cba8e74a2f3666a8083e1f6c6542))
+* tasks in dedicated file ([757793a](https://github.com/Stacktical/stacktical-developer-toolkit/commit/757793af15661af3264d99de6014fa6e9e1fb0d7))
+* tracking compose and job ([268c189](https://github.com/Stacktical/stacktical-developer-toolkit/commit/268c189b4a5155fd353440cf3745e5bccf5450c4))
+* **harmony:** deploying to harmonytestnet ([25919ad](https://github.com/Stacktical/stacktical-developer-toolkit/commit/25919ad44ca29c8a6ad6427484af31c2ad85b7a7))
+
+
+### Bug Fixes
+
+* set oracle permissions only if false ([67dc60f](https://github.com/Stacktical/stacktical-developer-toolkit/commit/67dc60f0eb1bba49b1e471e56c2b2cf78425f097))
+* SLARegistry deployment manual gas limit ([95c0399](https://github.com/Stacktical/stacktical-developer-toolkit/commit/95c0399338325e45598f7dcd8ec5b687df3cafb1))
+
 ## 1.0.0 (2021-05-21)